Normalize the vector in Transform#rotate()
Without normalization, using anything other than a unit vector
effectively serves to scale the rotation: a 15-degree rotation can
easily become 30 degrees (or more, or less...), for example.  This is, I
suspect, very confusing to newbies, and seeing as the main reason to
take a vector at all is to determine the rotation axis, better to
normalize it automatically, methinks.