Make kh2Bar demo work in Oozaru
Of course it's worth noting that this *barely* works--I had to change
the paths in the system modules as well as add a bunch of unimplemented
stub functions to `pegasus.ts`... but it *does* render the entire HP
gauge, so that's a start.  Now to work on actually implementing the
stubs!