Use ScriptProcessorNode instead of AudioBuffer
Now we're talking!  This gets me seamless playback of buffered audio...
with the downside that I can't control the sample rate.  So at the
moment, Chartreuse Rewind plays back a little on the slow side... which,
don't get me wrong, sounds awesome, but um... yeah, it's not supposed to
do that.  :o)