Resample using sine-based interpolation
It occurred to me that, while linear interpolation by itself doesn't
work well with sine-shaped data, sine is a function of angle--which *is*
linear.  This is the first time I've ever used `asin` for anything!

While I was doing this I also refactored the Stream class to fix an
issue where the amount of audio buffered wasn't reported correctly (it
reported more being buffered than actually was), causing the stream to
be starved.