Get rid of Sampler class, inline into Stream
`Sampler` was never meant to stick around, it was rather heavy and only
meant as an organizational tool to help me get a feel for how
intepolated sampling worked without being distracted by the buffering
logic.  Now that I have a better handle on it, the math can get inlined.