Pause streams when starved instead of skipping
When the Oozaru tab becomes inactive, requestAnimationFrame() stops
being honored and the event loop grinds to a halt.  This causes any
playing SoundStreams to be starved.  Previously, this caused the stream
to skip like a broken record, which was very unpleasant.  This change
makes the engine fill the output buffers with silence upon encountering
a buffer underrun, effectively pausing the stream--which is in line with
the API documentation included with miniSphere.