Add support for SoundStream#pause()
This was fairly easy to implement; since a ScriptProcessorNode is always
live, "pausing" it is best achieved by ignoring the buffered audio and
just sending silence until it becomes unpaused.

I suppose another way would have been to tear down the
ScriptProcessorNode while paused, but that seemed a bit heavy-handed.