Fix render-to-surface immediately after creation
When a Surface is first created, it's backing texture is created along
with it.  This leaves the graphics pipeline in a state where the
Surface's backing texture is selected both as a render source (bound to
texture unit 0) and as the render target (via its framebuffer).  WebGL
doesn't like this--at all--and it causes any and all draw calls to fail
until a different texture is bound to unit 0.

This change fixes the issue by not leaving a new texture bound after
creating it.  Instead, the original binding is remembered and restored
afterwards.