Tame Font#drawText() excessive CPU usage
Instead of uploading new vertices for each character in a line of text,
it ends up being much faster to just build a big triangle list, upload
it all at once, and issue a single draw call at the end.  This reduces
CPU usage and probably takes some of the load off the GPU too.