Fix WebGL warning when drawing empty string
Chrome produces annoying "render count or primcount is 0" warnings when
a draw call is issued asking to draw zero vertices.  This comes up in
practice when drawing an empty string; since nothing is going to be
drawn anyway, we can get rid of the warning and speed things up by just
returning from `drawText` early if passed an empty string.