Make Queue<T> into a deque
Now it's a complete double-ended queue implementation.  Everything is
amortized O(1), with one exception because I was lazy and didn't
consider the case common enough to worry about it: pop may end up
calling `Array#shift` in the case where an unshift has placed something
into the VIP stack.  This is an extreme case, as it first requires an
unshift to be done with a full deque (pushing the item onto the VIP
stack), followed by enough pops to drain both the overflow area *and*
the main ring.  Needless to say, I don't expect this case to come up
much in practice...