Add support for lazy-loading Font, Shader
Background loading of Shaders is a bit hacky; it needs to clone the
default shader to avoid clobbering uniforms, and any uniforms already
set will be lost once the real shader is swapped in.  Also I'm not sure
it's possible to make Shader#clone() work properly before the shader has
been fully loaded.