Improve error reporting
If an error is thrown out of a Dispatch job, the engine now catches it
and displays it to the user, as well as shutting down the Dispatch
machinery to induce a game crash.  Previously games would end up in a
kind of zombie state after an error because their Dispatch jobs would
keep being called as if nothing was wrong.