Add error for getting size of unloaded texture
Since new Texture() returns a 1x1 transparent image while loading the
real image in the background, this can throw off code that relies on
knowing the actual width and height of the texture image.  Instead of
having such code glitch out, this change makes Oozaru throw a RangeError
if the width or height are directly accessed before the texture has
fully loaded.  The game can use Texture.fromFile() if it needs
async-friendly loading.