Use VBO optimization for IBOs too
This worried me at first since I'm creating buffers on the GPU to be
rendered and then immediately thrown away, which seemed incredibly
inefficient.  But apparently, that's exactly the use case the "stream
draw" flag is intended for...

In any case, this greatly improves performance on Apple platforms
without taking on a hit on Windows and Linux (so far) so I'll take what
I can get.