Fix Fido percentage calculation, attempt 2
The last fix turned out to only partially fix the issue; there was still
an issue where if a fetch is started immediately after a previous one
finished, it wouldn't have time to check for completed jobs and end up
starting the new batch at a nonzero percentage.  This should fix that;
let's see if anything else goes wrong now.