Fix SoundStream buzzing with quick hack
Because the Audialis implementation stores the buffered input buffers
individually in a queue rather than concatenating them, the resampling
code ends up regularly reading past the end of an input buffer while
attempting to interpolate samples, causing NaNs to be written to the
output and generating a very loud, very annoying buzzing noise.  This
fixes it for now, by skipping interpolation if the end of an input
buffer is reached.  A more comprehensive fix will have to come later.