Disable synchronous asset loading
new Texture('filename'), etc. Anything that loads a file synchronously.

Oozaru used to emulate this by loading the file asynchronously and using
a placeholder resource in the meantime. This feature was always a bit
flaky. It worked well enough in basic cases, but a lot of code wants to
know the metrics of the thing it just loaded upfront, and using a
placeholder while the real thing loads in the background kind of gets in
the way of that. Better just to disable the constructors going forward.

The constructors will be deprecated in miniSphere as well. Once API
level 4 is finalized, they will be disabled for good.