Add API level check before launching game
If the game targets an API level not supported by the version of Oozaru
in use, Oozaru will throw an error instead of trying to launch it.

Also added a check, currently commented out, to prevent launching games
targeting API 3 or below.  This is a necessary step for Sphere platform
integration, since neoSphere currently supports APIs that are either
difficult or impossible to support on the Web and will need to be
disabled in games targeting API 4 or higher.