Add `Texture#whenReady` property
This one was tricky, not least of which because a rejected promise
without a catch handler is treated by JS as an uncaught exception.  I
managed to work around that by catching the exception and storing it for
later; if loading fails, a new rejected promise is returned later by the
`whenReady` getter.